pyproject.toml
[project]
name = "django-rest-framework"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asgiref==3.11.0",
    "attrs==25.4.0",
    "django==6.0.1",
    "django-filter==25.2",
    "django-silk==5.4.3",
    "djangorestframework==3.16.1",
    "djangorestframework-simplejwt==5.5.1",
    "drf-spectacular==0.29.0",
    "gprof2dot==2025.4.14",
    "inflection==0.5.1",
    "isort>=7.0.0",
    "jsonschema==4.26.0",
    "jsonschema-specifications==2025.9.1",
    "pillow==12.1.0",
    "pyjwt==2.11.0",
    "pyyaml==6.0.3",
    "referencing==0.37.0",
    "rpds-py==0.30.0",
    "sqlparse==0.5.5",
    "typing-extensions==4.15.0",
    "uritemplate==4.2.0",
]
